+++ dev/02-development.R
@@ -24,7 +24,7 @@
 #' List here all the command executed during development of the package
 #'
 
-# usethis::use_test("<function_name>")
+usethis::use_test("read_faers")
 usethis::use_r("read_faers")
 
 
@@ -45,6 +45,7 @@ usethis::use_r("read_faers")
 usethis::use_tidy_description()
 spelling::spell_check_package()
 spelling::update_wordlist()
+devtools::check_man()